[Pipelines] Filter pipelines for the project summary [1.8.x] (#7595)
(cherry picked from commit e5279ac36d1a4dbc122198c85538a839dacf1034)
